Clarify registration scoped of permission service
This commit will move the registration of class PermissionService to
an extension class.  This is intended to both secure and clarify that
the permission service needs to be registered as scoped, i.e. that it
needs to be created at every request.

If we did not, we might re-use the service instance from one user, for
another.  This is a catastrophic vulnerability.