Rename PermissionService
This commit will rename the PermissionService class to
HttpContextPersmissionService to better indicate that the
implementation depends on HttpContext and is specific to controller
logic.

This leave room for a ServiceBusPermissionService class, or any other
implementation of the IPermissionService interface that does not
originate from an HTTP call.

This commit also clarifies the comment in the AddPermissionService
extension method.