Update to .NET 7 and update tests
This commit will update to .NET 7 as well as updating tests in general
and system tests in particular.  The commit also fix a few other
household things, like removing the launchSettings.json files that was
added by the original templates.

This commit is probably a little too large, but you can think of it as
a modernization commit.

Co-authored-by: tobias.ahnoff@omegapoint.se