Add system test configuration using settings file
This commit will replace the TokenHttpClient class with a custom
delegate handler in combination with a base test class.

This simplifies the code and also means that we also where able to
implement a testsettings.json file that enables more real-world
scenarios such as running system tests on a CI server with
configuration.